+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## New functions:
 
+* `wav_2_flac()` to convert '.wav' files into flac compresed lossless format
 * `gaps()` for calculating time intervals (i.e. gaps) between selections
 * `read_sound_files()` is now use to read sound files
 * warbleR now supports '.wav', '.flac', '.wac' and '.mp3' formats